+++ forgerock-auth/src/test/java/org/forgerock/android/auth/FRUserMockTest.java
@@ -52,6 +52,7 @@ import java.util.concurrent.ExecutionException;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
@@ -265,6 +266,23 @@ public class FRUserMockTest extends BaseTest {
         FRListenerFuture<AccessToken> future = new FRListenerFuture<>();
         FRUser.getCurrentUser().getAccessToken(future);
         assertNotNull(future.get());
+    }
+
+    @Test
+    public void testRefreshTokenAsync() throws Exception {
+        frUserHappyPath();
+        FRListenerFuture<AccessToken> future = new FRListenerFuture<>();
+        FRUser.getCurrentUser().getAccessToken(future);
+        assertNotNull(future.get());
+        server.enqueue(new MockResponse().setResponseCode(HttpURLConnection.HTTP_OK));
+        enqueue("/authTreeMockTest_Authenticate_accessToken_no_RefreshToken_shortlife.json", HttpURLConnection.HTTP_OK);
+        //For revoke existing Access Token
+        FRListenerFuture<AccessToken> refreshTokenFuture = new FRListenerFuture<>();
+        FRUser.getCurrentUser().refresh(refreshTokenFuture);
+        server.takeRequest();
+        assertNotEquals(future.get().getExpiresIn(), refreshTokenFuture.get().getExpiresIn());
+        assertNotEquals(future.get().getValue(), refreshTokenFuture.get().getValue());
+        assertNotEquals(future.get().getIdToken(), refreshTokenFuture.get().getIdToken());
     }
 
     @Test(expected = AuthenticationException.class)