+++ forgerock-authenticator/src/test/java/org/forgerock/android/auth/PushFactoryTest.java
@@ -434,13 +434,14 @@ public class PushFactoryTest extends FRABaseTest {
                 "policies=eyJiaW9tZXRyaWNBdmFpbGFibGUiOiB7IH0sImRldmljZVRhbXBlcmluZyI6IHsic2NvcmUiOiAwLjh9fQ&" +
                 "digits=6&" +
                 "secret=R2PYFZRISXA5L25NVSSYK2RQ6E======&" +
-                "period=30&";
+                "period=30&" +
+                "issuer=Rm9yZ2VSb2Nr";
 
         factory.createFromUri(uri, pushListenerFuture);
         PushMechanism push = (PushMechanism) pushListenerFuture.get();
         assertEquals(push.getType(), Mechanism.PUSH);
         assertEquals(push.getAccountName(), "demo");
-        assertEquals(push.getIssuer(), "forgerock");
+        assertEquals(push.getIssuer(), "ForgeRock");
     }
 
 }