+++ samples/app/src/main/java/com/example/app/journey/Journey.kt
@@ -27,6 +27,7 @@ import com.example.app.callback.ConfirmationCallback
 import com.example.app.callback.DeviceBindingCallback
 import com.example.app.callback.DeviceProfileCallback
 import com.example.app.callback.DeviceSigningVerifierCallback
+import com.example.app.callback.IdPCallback
 import com.example.app.callback.NameCallback
 import com.example.app.callback.PasswordCallback
 import com.example.app.callback.PollingWaitCallback
@@ -41,6 +42,7 @@ import org.forgerock.android.auth.callback.ConfirmationCallback
 import org.forgerock.android.auth.callback.DeviceBindingCallback
 import org.forgerock.android.auth.callback.DeviceProfileCallback
 import org.forgerock.android.auth.callback.DeviceSigningVerifierCallback
+import org.forgerock.android.auth.callback.IdPCallback
 import org.forgerock.android.auth.callback.NameCallback
 import org.forgerock.android.auth.callback.PasswordCallback
 import org.forgerock.android.auth.callback.PollingWaitCallback
@@ -50,14 +52,16 @@ import org.forgerock.android.auth.callback.WebAuthnAuthenticationCallback
 import org.forgerock.android.auth.callback.WebAuthnRegistrationCallback
 
 @Composable
-fun Journey(journeyName: String,
-            journeyViewModel: JourneyViewModel, openDrawer: () -> Unit) {
+fun <T> Journey(journeyName: String,
+                journeyViewModel: JourneyViewModel<T>,
+                openDrawer: () -> Unit,
+                onCompletion: (() -> Unit)? = null) {
 
     val context = LocalContext.current
     val state by journeyViewModel.state.collectAsState()
 
     Journey(journeyName, state = state, openDrawer,
-        onNext = { state.node?.let { journeyViewModel.next(context, it) } })
+        onNext = { state.node?.let { journeyViewModel.next(context, it) } }, onCompletion)
 
 }
 
@@ -65,7 +69,8 @@ fun Journey(journeyName: String,
 fun Journey(journeyName: String,
             state: JourneyState,
             openDrawer: () -> Unit,
-            onNext: () -> Unit) {
+            onNext: () -> Unit,
+            onCompletion: (() -> Unit)?) {
 
     Column(modifier = Modifier
         .padding(8.dp)
@@ -73,14 +78,17 @@ fun Journey(journeyName: String,
         state.session?.apply {
             val userProfileViewModel =
                 viewModel<UserProfileViewModel>()
-            UserProfile(userProfileViewModel = userProfileViewModel, openDrawer)
+            onCompletion?.let { it() } ?: UserProfile(userProfileViewModel = userProfileViewModel, openDrawer)
         }
         state.exception?.apply {
             Error(exception = this, openDrawer)
+            onCompletion?.let { it() }
         }
         state.node?.apply {
             var showNext = true
-            Topbar(heading = "Journey - $journeyName", openDrawer = openDrawer)
+            if (journeyName.isNotBlank()) {
+                Topbar(heading = "Journey - $journeyName", openDrawer = openDrawer)
+            }
             state.node.callbacks?.forEach {
                 when (it) {
                     is NameCallback -> NameCallback(it)
@@ -89,29 +97,36 @@ fun Journey(journeyName: String,
                         DeviceBindingCallback(it, onCompleted = onNext)
                         showNext = false
                     }
+
                     is TextOutputCallback -> TextOutputCallback(it)
                     is DeviceSigningVerifierCallback -> {
                         DeviceSigningVerifierCallback(it, true, onCompleted = onNext)
                         showNext = false
                     }
+
                     is ConfirmationCallback -> {
                         ConfirmationCallback(it, onSelected = onNext)
                         showNext = false
                     }
+
                     is WebAuthnRegistrationCallback -> {
                         WebAuthnRegistrationCallback(it, state.node, onCompleted = onNext)
                         showNext = false
                     }
+
                     is WebAuthnAuthenticationCallback -> {
                         WebAuthnAuthenticationCallback(it, state.node, onCompleted = onNext)
                         showNext = false
                     }
+
                     is ChoiceCallback -> {
                         ChoiceCallback(it)
                     }
+
                     is PollingWaitCallback -> {
                         PollingWaitCallback(it, onTimeout = onNext)
                     }
+
                     is SelectIdPCallback -> {
                         SelectIdPCallback(callback = it, onSelected = onNext)
                     }
@@ -123,6 +138,10 @@ fun Journey(journeyName: String,
                      */
                     is DeviceProfileCallback -> {
                         DeviceProfileCallback(callback = it, onCompleted = onNext)
+                        showNext = false
+                    }
+                    is IdPCallback -> {
+                        IdPCallback(callback = it, onCompleted = onNext)
                         showNext = false
                     }